# job-platform/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MedMatch Job Platform API",
    "description": "API for job matching platform connecting medical professionals with healthcare employers. Includes application tracking, dashboard analytics, and notification system.",
    "version": "1.1.0",
    "contact": {
      "name": "MedMatch API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/notifications": {
      "get": {
        "summary": "List notifications",
        "description": "Get notifications for a user",
        "tags": ["Notifications"],
        "parameters": [
          {
            "name": "userType",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "enum": ["candidate", "employer"] },
            "description": "Type of user"
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" },
            "description": "User ID"
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": { "type": "string", "enum": ["true", "false"] },
            "description": "If true, only return unread notifications"
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Notification" }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "unreadCount": { "type": "integer" }
                      }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing or invalid parameters" }
        }
      },
      "delete": {
        "summary": "Delete read notifications",
        "description": "Delete all read notifications for a user",
        "tags": ["Notifications"],
        "parameters": [
          {
            "name": "userType",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "enum": ["candidate", "employer"] }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Notifications deleted successfully" }
        }
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "summary": "Mark notification as read",
        "description": "Mark a specific notification as read",
        "tags": ["Notifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": { "description": "Notification marked as read" },
          "404": { "description": "Notification not found" }
        }
      }
    },
    "/notifications/mark-all-read": {
      "post": {
        "summary": "Mark all notifications as read",
        "description": "Mark all notifications as read for a user",
        "tags": ["Notifications"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userType", "userId"],
                "properties": {
                  "userType": { "type": "string", "enum": ["candidate", "employer"] },
                  "userId": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "All notifications marked as read" },
          "400": { "description": "Missing parameters" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userType": { "type": "string", "enum": ["candidate", "employer"] },
          "userId": { "type": "integer" },
          "type": { "type": "string" },
          "message": { "type": "string" },
          "read": { "type": "integer", "enum": [0, 1] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "hasNextPage": { "type": "boolean" },
          "hasPrevPage": { "type": "boolean" }
        }
      }
    }
  }
}